refactor(audit): split monolithic audit.go into separate focused components
- Extract logger functionality into logger.go with dedicated options
- Extract reader functionality into reader.go
- Move event creation options to event_options.go
- Improve separation of concerns with dedicated files for each component
- Remove deprecated audit.go and context.go files
- Maintain backwards compatibility with same public API
